hero.h, hero.cpp, heroState.txt
This update includes a loadState() function that loads the hero's previous state from the heroState.txt file, a saveState() function that saves the hero's current state, a isAlive() function that returns true if the hero is still alive, and false if the hero is dead, and a resetState() function, that resets the heroState.txt to the hero's original whiffles: 0, energy: 100 and x: 11, y: 13. Wiki page will be updated momentarily.